+++ crates/vm/levm/src/opcode_handlers/system.rs
@@ -768,14 +768,14 @@ impl<'a> VM<'a> {
 
             // Return gas left from subcontext
             if ctx_result.is_success() {
-                call_frame.gas_remaining = call_frame
-                    .gas_remaining
+                call_frame.gas_remaining = (call_frame.gas_remaining as u64)
                     .checked_add(
                         gas_limit
                             .checked_sub(ctx_result.gas_used)
                             .ok_or(InternalError::Underflow)?,
                     )
-                    .ok_or(InternalError::Overflow)?;
+                    .ok_or(InternalError::Overflow)?
+                    as i64;
             }
 
             // Store return data of sub-context
@@ -894,7 +894,7 @@ impl<'a> VM<'a> {
             .ok_or(InternalError::Underflow)?;
         parent_call_frame.gas_remaining = parent_call_frame
             .gas_remaining
-            .checked_add(child_unused_gas)
+            .checked_add(child_unused_gas as i64)
             .ok_or(InternalError::Overflow)?;
 
         // Store return data of sub-context
@@ -955,7 +955,7 @@ impl<'a> VM<'a> {
             .ok_or(InternalError::Underflow)?;
         parent_call_frame.gas_remaining = parent_call_frame
             .gas_remaining
-            .checked_add(unused_gas)
+            .checked_add(unused_gas as i64)
             .ok_or(InternalError::Overflow)?;
 
         // What to do, depending on TxResult
@@ -1006,12 +1006,12 @@ impl<'a> VM<'a> {
         let gas_left = self
             .current_call_frame
             .gas_remaining
-            .checked_sub(eip7702_gas_consumed)
+            .checked_sub(eip7702_gas_consumed as i64)
             .ok_or(ExceptionalHalt::OutOfGas)?;
 
         Ok((
             new_memory_size,
-            gas_left,
+            gas_left as u64,
             account_is_empty,
             address_was_cold,
         ))
@@ -1027,7 +1027,7 @@ impl<'a> VM<'a> {
         // Return gas_limit to callframe.
         callframe.gas_remaining = callframe
             .gas_remaining
-            .checked_add(gas_limit)
+            .checked_add(gas_limit as i64)
             .ok_or(InternalError::Overflow)?;
         callframe.stack.push1(FAIL)?; // It's the same as revert for CREATE
